Pulling PostFormVerification out to separate class
